+++ src/funcs/sum.h
@@ -5,7 +5,7 @@
 
 #include "../core/func.h"
 
-double sum_eval(unsigned long index, int rate, Func **args, int count, __attribute__((unused)) void *context)
+double sum_eval(unsigned long index, double rate, Func **args, int count, __attribute__((unused)) void *context)
 {
     double sum = 0;
     for (int i = 0; i < count; i++)
